Day5_DBs/pyproject.toml
[tool.poetry]
name = "day5-dbs"
version = "0.1.0"
description = "learning or practicing database postgresql with fastapi"
authors = ["sharyark <ksharyar12345@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.3"
uvicorn = "^0.29.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
